NRPT:297. Clean out PermitAmendment, migration
Initial cleanup of the PermitAmendment and PermitAmendmentBCMI flavour. Update for PermitBCMI model. Creation of a migration script to flatten the PermitAmendment model and delete old objects.
