--- api/src/controllers/put/permit-amendment.js
@@ -1,98 +1,0 @@
-const mongoose = require('mongoose');
-const ObjectID = require('mongodb').ObjectID;
-const PutUtils = require('../../utils/put-utils');
-const PermitPost = require('../post/permit');
-
-/**
- * Performs all operations necessary to edit a master Permit Amendment record and its associated flavour records.
- *
- * @param {*} args
- * @param {*} res
- * @param {*} next
- * @param {*} incomingObj see example
- * @returns object containing the operation's status and created records
- */
-exports.editRecord = async function (args, res, next, incomingObj, overridePutParams = null) {
-  const flavourFunctions = {
-    PermitAmendmentBCMI: this.editBCMI
-  }
-  return await PutUtils.editRecordWithFlavours(args, res, next, incomingObj, this.editMaster, PermitPost, 'PermitAmendment', flavourFunctions, overridePutParams);
-};
-
-
-/**
- * Performs all operations necessary to edit a master Permit Amendment record.
- * 
- * @param {*} args
- * @param {*} res
- * @param {*} next
- * @param {*} incomingObj see example
- * @returns edited master permit amendment record
- */
-exports.editMaster = function (args, res, next, incomingObj, flavourIds) {
-  delete incomingObj._id;
-
-  // Reject any changes to master permissions
-  delete incomingObj.read;
-  delete incomingObj.write;
-
-  const PermitAmendment = mongoose.model('PermitAmendment');
-  const sanitizedObj = PutUtils.validateObjectAgainstModel(PermitAmendment, incomingObj);
-  if (!sanitizedObj || sanitizedObj === {}) {
-    // skip, as there are no changes to master record
-    return;
-  }
-
-  sanitizedObj.dateUpdated = new Date();
-  sanitizedObj.updatedBy = args.swagger.params.auth_payload.displayName;
-
-  const dotNotatedObj = PutUtils.getDotNotation(sanitizedObj);
-  const updateObj = { $set: dotNotatedObj };
-
-  if (flavourIds && flavourIds.length) {
-    updateObj.$addToSet = { _flavourRecords: flavourIds.map(id => new ObjectID(id)) };
-  }
-
-  return updateObj;
-};
-
-/**
- * Performs all operations necessary to edit a lng Permit record.
- *
- * @param {*} args
- * @param {*} res
- * @param {*} next
- * @param {*} incomingObj see example
- * @returns edited BCMI permit amendment record
- */
-exports.editBCMI = function (args, res, next, incomingObj) {
-  delete incomingObj._id;
-
-  // Reject any changes to permissions
-  // Publishing must be done via addRole or removeRole
-  delete incomingObj.read;
-  delete incomingObj.write;
-
-  let PermitBCMI = mongoose.model('PermitAmendmentBCMI');
-
-  const sanitizedObj = PutUtils.validateObjectAgainstModel(PermitBCMI, incomingObj);
-
-  sanitizedObj.dateUpdated = new Date();
-
-  const dotNotatedObj = PutUtils.getDotNotation(sanitizedObj);
-
-  // If incoming object has addRole: 'public' then read will look like ['sysadmin', 'public']
-  const updateObj = { $set: dotNotatedObj,  $addToSet: {}, $pull: {} };
-
-  if (incomingObj.addRole && incomingObj.addRole === 'public') {
-    updateObj.$addToSet['read'] = 'public';
-    updateObj.$set['datePublished'] = new Date();
-    updateObj.$set['publishedBy'] = args.swagger.params.auth_payload.displayName;
-  } else if (incomingObj.removeRole && incomingObj.removeRole === 'public') {
-    updateObj.$pull['read'] = 'public';
-    updateObj.$set['datePublished'] = null;
-    updateObj.$set['publishedBy'] = '';
-  }
-
-  return updateObj;
-};
