+++ api/src/models/bcmi/index.js
@@ -1,4 +1,3 @@
 exports.mineBCMI = require('./mine-bcmi');
 exports.permitBCMI = require('./permit-bcmi');
-exports.permitAmendmentBCMI = require('./permitAmendment-bcmi');
 exports.collectionBCMI = require('./collection-bcmi');
